Working Version: MongoDriver works
Problem: Our Java function blocks the server -> The driver never gets
data -> Starvation
Solution: In C++ we call the java in a background thread, so the client
gets a dummy result, bit the server can process run without any problem.
This is enough for our demo, since now we have access to the db from
java and we can read/store everything from/in the db

One current problem: after the 2. run it crasches.